Code refactoring and test update
