+++ Car/__tests__/activity/journey/OkSearchResult-test.tsx
@@ -1,13 +1,24 @@
 import React from "react";
 import shallowRender from "react-test-renderer/shallow";
 import OkSearchResult from "../../../src/activity/journey/journey-activity/search-journey/search-results/ok-search-result/OkSearchResult";
+import OkSearchResultProps from "../../../src/activity/journey/journey-activity/search-journey/search-results/ok-search-result/OkSearchResultProps";
 
 const renderer = shallowRender.createRenderer();
+
+const props: OkSearchResultProps = {
+    route: {
+        params: {
+            journeys: [],
+            displayFee: false,
+            passangersCount: 1
+        },
+    },
+};
 
 test("renders correctly", async () =>
     expect(
         renderer.render(
-            <OkSearchResult journeys={[]} displayFee={false} passangersCount={1} />
+            <OkSearchResult {...props} />
         )
     ).toMatchInlineSnapshot(`
     <View
