+++ Car/src/activity/journey/journey-activity/search-journey/SearchJourneyStyle.tsx
@@ -11,12 +11,12 @@ const SearchJourneyStyle = StyleSheet.create({
         paddingLeft: 16,
         paddingRight: 15
     },
-    
+
     DepartureContainerLocation: {
         paddingLeft: 16,
         paddingRight: 15
     },
-    
+
     viaContainerLocation: {
         paddingLeft: 16,
         paddingRight: 15
