+++ Car/src/activity/journey/journey-activity/search-journey/search-results/ok-search-result/OkSearchResultProps.tsx
@@ -1,0 +1,14 @@
+import ApplicantJourney from "../../../../../../../models/journey/ApplicantJourney";
+
+interface OkSearchResultProps
+{
+    route: {
+        params: {
+            journeys: ApplicantJourney[],
+            displayFee: boolean,
+            passangersCount: number
+        }
+    }
+}
+
+export default OkSearchResultProps;
